# set cpubw perf governor