@geohot 2020-11-18
@adeebshihadeh Do we have a test for offroad power? That's my only concern with this.