@adeebshihadeh 2020-11-18
Not yet, but I wouldn't be too worried about offroad power usage increasing significantly with this change.