@geohot 2020-11-18
Oh haha, this is already set in thermald!